Python/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
colorama==0.4.6
confluent-kafka==2.2.0
crashtest==0.4.1
cryptography==41.0.3
distlib==0.3.7
dulwich==0.21.5
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.0.0
msgpack==1.0.5
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pexpect==4.8.0
pipenv==2023.7.23
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
rapidfuzz==2.15.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.10.1
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.20
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.7.6
virtualenv==20.24.2
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.3.7
zipp==3.16.2
